统一Long为数字格式

+++ acgbox-content/src/main/java/org/killze/acgbox/service/AdaptationTypeService.java
@@ -6,7 +6,7 @@ import org.killze.acgbox.vo.content.AdaptationTypeVO;
 import java.util.List;
 
 /**
- * 改编类型服务接口。
+ * 改编类型服务接口
  *
  * @author killze
  */
@@ -15,35 +15,33 @@ public interface AdaptationTypeService {
     /**
      * 创建改编类型
      *
-     * @param dto 改编类型数据传输对象
-     * @return 创建成功的改编类型信息
+     * @param dto 改编类型信息
+     * @return 改编类型信息
      */
     AdaptationTypeVO createAdaptationType(AdaptationTypeDTO dto);
 
     /**
      * 修改改编类型
      *
-     * @param dto 改编类型数据传输对象
-     * @return 修改成功的改编类型信息
+     * @param dto 改编类型信息
+     * @return 改编类型信息
      */
     AdaptationTypeVO updateAdaptationType(AdaptationTypeDTO dto);
 
     /**
      * 批量删除改编类型
      *
-     * @param ids 改编类型ID列表
+     * @param ids 改编类型 ID 列表
      */
-    void deleteAdaptationTypes(List<Integer> ids);
+    void deleteAdaptationTypes(List<Long> ids);
 
     /**
-     * 根据ID获取改编类型
+     * 根据 ID 获取改编类型
      *
-     * @param id 改编类型ID
+     * @param id 改编类型 ID
      * @return 改编类型信息
      */
-    AdaptationTypeVO getAdaptationTypeById(
-            Integer id
-    );
+    AdaptationTypeVO getAdaptationTypeById(Long id);
 
     /**
      * 获取全部改编类型
